clinker-core: extract Route, Merge, Sort, CorrelationCommit arms to focused modules
Move four PlanNode arm bodies out of executor/dispatch.rs into new sibling
modules:

- route_dispatch.rs   — branch predicate evaluation, port emission, default
                        handling.
- merge_dispatch.rs    — concat / interleave concatenation plus the fused
                        all-Source crossbeam-Select interleave fast path.
- sort_dispatch.rs     — planner-synthesized enforcer-sort buffer + spill.
- correlation_dispatch.rs — the CorrelationCommit arm delegator plus the
                        per-group correlation-buffer commit machinery
                        (commit_correlation_buffers, format_group_key,
                        commit_one_group, flush_clean_records_to_writers).

Each dispatcher match arm is now a single delegating call. The combine
buffer entry types (CorrelationGroupBuffer / CorrelationRecordSlot /
CorrelationErrorRecord) stay in dispatch.rs because they are part of
ExecutorContext's surface and the shared error-routing path; the commit
subtree still reaches commit_correlation_buffers through the
crate::executor::dispatch path via a re-export.

Pure relocation: no symbol renames, no public-API change.

Closes #247
