+++ crates/clinker-core/src/executor/dispatch.rs
@@ -29,6 +29,11 @@ use petgraph::graph::NodeIndex;
 use crate::config::{ErrorStrategy, OutputConfig, PipelineConfig};
 use crate::error::PipelineError;
 
+/// Re-export of the correlation-buffer commit entry point, relocated to
+/// [`crate::executor::correlation_dispatch`]. The commit subtree reaches it
+/// through this module path, so the re-export preserves that surface.
+pub(crate) use crate::executor::correlation_dispatch::commit_correlation_buffers;
+
 /// Stand-in `$source.file` value for dispatch sites that have no
 /// originating source record (Combine/finalize/post-aggregate emits with
 /// `source_row: 0`, body composition emits that lost lineage at fan-in).
@@ -408,7 +413,7 @@ use crate::executor::node_buffer::NodeBuffer;
 use crate::executor::schema_check::check_input_schema;
 use crate::executor::{
     CompiledRoute, CompiledTransform, DlqEntry, build_format_writer, evaluate_single_transform,
-    evaluate_single_transform_windowed, parse_memory_limit, stage_metrics,
+    evaluate_single_transform_windowed, stage_metrics,
 };
 use crate::pipeline::memory::BudgetCategory;
 use crate::plan::bind_schema::CompileArtifacts;
@@ -2921,796 +2926,16 @@ pub(crate) fn dispatch_plan_node(
             ctx.node_buffers.insert(node_idx, nb);
         }
 
-        PlanNode::Route {
-            ref name,
-            mode,
-            branches: _,
-            default: _,
-            ..
-        } => {
-            // Body-context Routes that consume an input port have no
-            // predecessor in the body's mini-DAG — the records are
-            // seeded into this node's own buffer at composition entry.
-            // Check own buffer first, fall back to predecessor.
-            let predecessors: Vec<NodeIndex> = current_dag
-                .graph
-                .neighbors_directed(node_idx, Direction::Incoming)
-                .collect();
-            let (input_records, input_puncts): (
-                Vec<(Record, u64)>,
-                Vec<crate::executor::stream_event::Punctuation>,
-            ) = if let Some(own_buf) = drain_node_buffer_slot(ctx, node_idx) {
-                own_buf.drain_split()?
-            } else {
-                match predecessors
-                    .iter()
-                    .find_map(|p| drain_node_buffer_slot(ctx, *p))
-                {
-                    Some(nb) => nb.drain_split()?,
-                    None => (Vec::new(), Vec::new()),
-                }
-            };
-
-            if let Some(expected) = current_dag.graph[node_idx]
-                .expected_input_schema_in(current_dag)
-                .cloned()
-            {
-                let upstream_name = predecessors
-                    .first()
-                    .map(|&i| current_dag.graph[i].name().to_string())
-                    .unwrap_or_default();
-                for (record, _) in &input_records {
-                    check_input_schema(&expected, record.schema(), name, "route", &upstream_name)?;
-                }
-            }
-
-            // Get successor nodes (branch transform nodes)
-            let successors: Vec<NodeIndex> = current_dag
-                .graph
-                .neighbors_directed(node_idx, Direction::Outgoing)
-                .collect();
-
-            // Build branch_name -> successor NodeIndex mapping.
-            //
-            // Under the unified taxonomy, a
-            // `PlanNode::Route` is a first-class node named as the
-            // user wrote it in YAML (e.g. `classify_route`). Successors
-            // may be Transforms, Aggregates, Outputs, or Compositions —
-            // any node whose `input:` is declared as
-            // `<route_name>.<branch>` (NodeInput::Port form). We
-            // can't use `build_transform_specs` alone because it
-            // excludes Outputs and Compositions.
-            //
-            // Walk `config.nodes` once, read each node's
-            // `header.input`, match against `<route_parent>.<branch>`,
-            // and map the branch to the successor's NodeIndex.
-            let route_parent = name.strip_prefix("route_").unwrap_or(name);
-            let mut succ_by_name: HashMap<String, NodeIndex> = HashMap::new();
-            for &succ in &successors {
-                succ_by_name.insert(current_dag.graph[succ].name().to_string(), succ);
-            }
-            let mut branch_to_succ: HashMap<String, NodeIndex> = HashMap::new();
-            // Iterate either the body's input-ref table (when set
-            // by the body executor for the current walk) or the
-            // top-level config nodes. The shape is the same:
-            // `(consumer_name, input_ref)` pairs.
-            let mut name_input_pairs: Vec<(String, String)> = Vec::new();
-            if let Some(body_refs) = ctx.current_body_node_input_refs.as_ref() {
-                for (consumer, refs) in body_refs {
-                    for r in refs {
-                        name_input_pairs.push((consumer.clone(), r.clone()));
-                    }
-                }
-            } else {
-                use crate::config::PipelineNode;
-                use crate::config::node_header::NodeInput;
-                for spanned in &ctx.config.nodes {
-                    let (node_name, input_ref) = match &spanned.value {
-                        PipelineNode::Transform { header, .. }
-                        | PipelineNode::Aggregate { header, .. }
-                        | PipelineNode::Route { header, .. }
-                        | PipelineNode::Output { header, .. } => {
-                            let ir = match &header.input.value {
-                                NodeInput::Single(s) => s.clone(),
-                                NodeInput::Port { node, port } => format!("{node}.{port}"),
-                            };
-                            (header.name.clone(), ir)
-                        }
-                        _ => continue,
-                    };
-                    name_input_pairs.push((node_name, input_ref));
-                }
-            }
-            for (node_name, input_ref) in name_input_pairs {
-                // Port form: "<route>.<branch>" — branch name comes
-                // from the port suffix. Standard for named branches
-                // on routes with multiple downstream consumers per
-                // branch.
-                if let Some(branch) = input_ref.strip_prefix(&format!("{}.", route_parent))
-                    && let Some(&succ) = succ_by_name.get(&node_name)
-                {
-                    branch_to_succ.insert(branch.to_string(), succ);
-                    continue;
-                }
-                // Bare form: `input: <route>` on an Output whose
-                // *name* matches a Route branch (or the Route
-                // `default:`). Conventional shorthand when each
-                // branch has exactly one downstream consumer and
-                // the consumer's name is the branch name.
-                if input_ref == route_parent
-                    && let Some(&succ) = succ_by_name.get(&node_name)
-                {
-                    branch_to_succ.insert(node_name.clone(), succ);
-                }
-            }
-
-            // Initialize per-successor buffers
-            let mut branch_buffers: HashMap<NodeIndex, Vec<_>> = HashMap::new();
-            for &succ in &successors {
-                branch_buffers.insert(succ, Vec::new());
-            }
-
-            // Per-route compiled evaluator wins over the singleton
-            // — `compiled_routes_by_name` is populated for body
-            // Routes (and any Route whose conditions need an explicit
-            // lookup), while the singleton remains the long-standing
-            // path for the top-level Route. A body's Route arm
-            // therefore finds its own conditions here instead of
-            // inheriting the top-level Route's conditions, which
-            // would have been the only ones available before.
-            // `take` + restore keeps the route's `&mut`-mutated state
-            // (regex caches, evaluator counters) consistent across
-            // records hitting the same Route node within one walk.
-            let from_map = ctx.compiled_routes_by_name.remove(name.as_str());
-            let mut from_singleton_flag = false;
-            let mut route_handle = match from_map {
-                Some(r) => Some(r),
-                None => {
-                    let taken = ctx.compiled_route.take();
-                    from_singleton_flag = taken.is_some();
-                    taken
-                }
-            };
-
-            if let Some(ref mut route) = route_handle {
-                for (i, (record, rn)) in input_records.into_iter().enumerate() {
-                    // Poll the shutdown flag every 1024 records so a long
-                    // Route chain terminates promptly on SIGINT. Same
-                    // chunk-boundary cadence the Transform arm uses.
-                    if i > 0 && i.is_multiple_of(1024) {
-                        ctx.check_shutdown()?;
-                    }
-                    let source_file_arc = source_file_arc_of(&record);
-                    let source_name_arc = source_name_arc_of(&record);
-                    let eval_ctx = EvalContext {
-                        stable: ctx.stable,
-                        source_file: &source_file_arc,
-                        source_row: rn,
-                        source_path: &source_file_arc,
-                        source_count: ctx.source_count_by_name(&source_name_arc),
-                        source_batch: ctx.source_batch_arc,
-                        ingestion_timestamp: ctx.source_ingestion_timestamp,
-                        source_name: &source_name_arc,
-                        doc_ctx: record.doc_ctx(),
-                    };
-
-                    let route_result = {
-                        let _guard = ctx.route_timer.guard();
-                        route.evaluate(&record, &eval_ctx)
-                    };
-                    match route_result {
-                        Ok(targets) => {
-                            for target in &targets {
-                                if let Some(&succ) = branch_to_succ.get(target.as_str()) {
-                                    branch_buffers
-                                        .entry(succ)
-                                        .or_default()
-                                        .push((record.clone(), rn));
-                                }
-                                // Exclusive mode: stop after first match
-                                if mode == crate::config::RouteMode::Exclusive {
-                                    break;
-                                }
-                            }
-                            advance_cursor(ctx, &source_name_arc, rn);
-                        }
-                        Err(route_err) => {
-                            if ctx.strategy == ErrorStrategy::FailFast {
-                                return Err(route_err.into());
-                            }
-                            let stage = Some(DlqEntry::stage_route_eval());
-                            let routed = record_error_to_buffer_if_grouped(
-                                ctx,
-                                &record,
-                                rn,
-                                crate::dlq::DlqErrorCategory::TypeCoercionFailure,
-                                route_err.to_string(),
-                                stage.clone(),
-                                None,
-                            );
-                            if !routed {
-                                let source_name = source_name_arc_of(&record);
-                                let triggering_field = route_err.triggering_field.clone();
-                                let triggering_value = route_err.triggering_value();
-                                push_dlq(
-                                    ctx,
-                                    DlqEntry {
-                                        source_row: rn,
-                                        category: crate::dlq::DlqErrorCategory::TypeCoercionFailure,
-                                        error_message: route_err.to_string(),
-                                        original_record: record,
-                                        stage,
-                                        route: None,
-                                        trigger: true,
-                                        source_name,
-                                        triggering_field,
-                                        triggering_value,
-                                    },
-                                )?;
-                            }
-                        }
-                    }
-                }
-            }
-            // Restore the route to whichever storage it came from.
-            if let Some(r) = route_handle {
-                if from_singleton_flag {
-                    ctx.compiled_route = Some(r);
-                } else {
-                    ctx.compiled_routes_by_name.insert(name.to_string(), r);
-                }
-            }
-
-            // Streaming-Output handoff: a single-branch Route whose sole
-            // successor is a streaming-eligible Output installed its
-            // sender under our `node_idx` at executor entry. The one
-            // branch's records are already collected into `merged`, so
-            // this does not shrink the Route's own working set; what it
-            // saves is the second copy — the records stream straight to
-            // the writer thread rather than crossing a charged
-            // `node_buffers` slot the Output would re-drain, and the
-            // writer overlaps with the next topo node. The eligibility
-            // predicate certified the single outgoing edge, so
-            // `branch_buffers` holds exactly the one successor here, and a
-            // streaming Route → terminal Output crosses no deferred
-            // region, so the cross-region tee below is correctly skipped.
-            // Dropping the sender disconnects the writer's recv.
-            if let Some(sender) = ctx.take_streaming_sender(node_idx) {
-                let batch_size = ctx.batch_size;
-                let spill_allowed = node_buffer_spill_allowed(current_dag, node_idx);
-                let charge = ctx
-                    .streaming_charge_handle(node_idx, name, spill_allowed)
-                    .expect("streaming sender implies a registered charge consumer");
-                let merged: Vec<(Record, u64)> = branch_buffers.into_values().flatten().collect();
-                stream_linear_producer_emit(
-                    &sender,
-                    batch_size,
-                    name,
-                    merged,
-                    input_puncts,
-                    &charge,
-                )?;
-                return Ok(());
-            }
-
-            // Put branch buffers into node_buffers keyed by successor.
-            // For successors that fall inside a deferred region while
-            // this Route does not, also park the per-branch records on
-            // the matching outgoing edge so the commit-time deferred
-            // dispatcher receives the same records the forward branch
-            // assignment selected. Internal-region edges and edges
-            // between two non-deferred operators skip the tee — the
-            // node_buffers entry already covers them.
-            let route_region_producer =
-                current_dag.deferred_region_at(node_idx).map(|r| r.producer);
-            let active_body = ctx.window_runtime.active_stack.last().copied();
-            for (succ_idx, buf) in branch_buffers {
-                let succ_region_producer =
-                    current_dag.deferred_region_at(succ_idx).map(|r| r.producer);
-                let crosses = match (route_region_producer, succ_region_producer) {
-                    (None, Some(_)) => true,
-                    (Some(p), Some(t)) if p != t => true,
-                    _ => false,
-                };
-                if crosses && let Some(edge) = current_dag.graph.find_edge(node_idx, succ_idx) {
-                    let row_bytes_each: u64 = buf
-                        .first()
-                        .map(|(rec, _)| {
-                            (std::mem::size_of::<Value>() * rec.schema().column_count()
-                                + std::mem::size_of::<(Record, u64)>())
-                                as u64
-                        })
-                        .unwrap_or(0);
-                    for (record, rn) in &buf {
-                        if row_bytes_each > 0 && ctx.memory_budget.should_abort() {
-                            return Err(PipelineError::MemoryBudgetExceeded {
-                                node: name.to_string(),
-                                used: ctx.memory_budget.peak_rss().unwrap_or(0),
-                                limit: ctx.memory_budget.hard_limit(),
-                                source: BudgetCategory::Arena,
-                                detail: Some("Route cross-region tee admission".to_string()),
-                            });
-                        }
-                        ctx.region_input_buffers
-                            .entry((active_body, edge))
-                            .or_default()
-                            .push((record.clone(), *rn));
-                    }
-                }
-                // Route broadcasts punctuations to every branch — each
-                // downstream subgraph needs the same document-boundary
-                // signal to drive its own per-document accumulators
-                // and writers.
-                let nb = admit_node_buffer(
-                    ctx,
-                    name,
-                    succ_idx,
-                    buf,
-                    input_puncts.clone(),
-                    node_buffer_spill_allowed(current_dag, succ_idx),
-                )?;
-                ctx.node_buffers.insert(succ_idx, nb);
-            }
+        PlanNode::Route { .. } => {
+            crate::executor::route_dispatch::dispatch_route(ctx, current_dag, node_idx, &node)?;
         }
 
-        PlanNode::Merge { ref name, .. } => {
-            // Two architectural modes for a Merge arm:
-            //
-            // 1. **Fused** — every direct predecessor is a Source and
-            //    `mode: interleave`. The pre-pass at executor entry
-            //    marked each predecessor's source name in
-            //    `ctx.fused_sources`, the Source dispatch arms
-            //    returned cleanly without consuming, and the receivers
-            //    are still parked in `ctx.source_records`. This arm
-            //    takes ownership of those receivers and runs a fair
-            //    `crossbeam_channel::Select` over them so a slow Source
-            //    no longer blocks peers — back-pressure flows
-            //    end-to-end.
-            //
-            // 2. **Non-fused** — concat mode, or a mix of Source and
-            //    non-Source predecessors. Predecessor arms have
-            //    already populated `ctx.node_buffers`; this arm
-            //    consumes those buffers in declaration order (Concat)
-            //    or round-robins across them (Interleave).
-            //
-            // Per-source FIFO survives both modes; per-record schema
-            // canonicalization runs at consume time so the canonical
-            // `Arc<Schema>` reaches downstream operators uniformly.
-            let predecessors: Vec<NodeIndex> = current_dag
-                .graph
-                .neighbors_directed(node_idx, Direction::Incoming)
-                .collect();
-
-            let (declaration_order, mode, interleave_seed): (
-                Vec<String>,
-                crate::config::MergeMode,
-                Option<u64>,
-            ) = {
-                use crate::config::PipelineNode;
-                use crate::config::node_header::NodeInput;
-                ctx.config
-                    .nodes
-                    .iter()
-                    .find_map(|spanned| match &spanned.value {
-                        PipelineNode::Merge { header, config, .. } if header.name == *name => {
-                            let order = header
-                                .inputs
-                                .iter()
-                                .map(|ni| match &ni.value {
-                                    NodeInput::Single(s) => s.clone(),
-                                    NodeInput::Port { node, port } => {
-                                        format!("{node}.{port}")
-                                    }
-                                })
-                                .collect();
-                            Some((order, config.mode, config.interleave_seed))
-                        }
-                        _ => None,
-                    })
-                    .unwrap_or_else(|| (Vec::new(), crate::config::MergeMode::Concat, None))
-            };
-
-            // Sort predecessors by declaration order
-            let mut sorted_preds = predecessors.clone();
-            sorted_preds.sort_by_key(|p| {
-                let pred_name = current_dag.graph[*p].name();
-                declaration_order
-                    .iter()
-                    .position(|d| d == pred_name)
-                    .unwrap_or(usize::MAX)
-            });
-
-            let merge_output_schema = current_dag.graph[node_idx].stored_output_schema().cloned();
-            // Detect fused mode: every predecessor is a Source whose
-            // name is in `ctx.fused_sources`. The pre-pass at executor
-            // entry already validated mode == Interleave for these,
-            // but we double-check here defensively.
-            let fused_mode = matches!(mode, crate::config::MergeMode::Interleave)
-                && sorted_preds.iter().all(|p| match &current_dag.graph[*p] {
-                    PlanNode::Source { name: src_name, .. } => {
-                        ctx.fused_sources.contains(src_name.as_str())
-                    }
-                    _ => false,
-                });
-
-            // Punctuation collection state hoisted above the if-else so
-            // the dedup pass below the merge body can read it. The
-            // fused-mode merge_fused_interleave path does not yet
-            // forward punctuations (a follow-up will route them
-            // through merge_fused_interleave's recv loop); the else
-            // branch's drain logic fills `all_puncts` from each input
-            // NodeBuffer.
-            let fan_in_degree = sorted_preds.len();
-            let mut all_puncts: Vec<crate::executor::stream_event::Punctuation> = Vec::new();
-
-            // Streaming-Output handoff: if a single Output downstream of
-            // this Merge passed the eligibility predicate at executor
-            // entry, its crossbeam `Sender` was installed under our
-            // `node_idx`. Take it here so the Merge arm streams every
-            // record through the channel instead of accumulating, and so
-            // dropping the sender at clean exit disconnects the streaming
-            // thread's `recv` loop. The fused-interleave path streams
-            // inside `merge_fused_interleave` (the sender moves there);
-            // the non-fused path (concat, or interleave with non-Source
-            // inputs) accumulates `merged` then streams it through
-            // `stream_linear_producer_emit` below, skipping the
-            // materialized `admit_node_buffer` slot.
-            let streaming_sender = ctx.take_streaming_sender(node_idx);
-            // The per-batch charge handle for the streaming slot, shared by
-            // both fused and non-fused streaming paths. `None` when this
-            // Merge is materialized (no streaming sender installed).
-            let merge_charge = streaming_sender.as_ref().and_then(|_| {
-                let spill_allowed = node_buffer_spill_allowed(current_dag, node_idx);
-                ctx.streaming_charge_handle(node_idx, name, spill_allowed)
-            });
-            let merge_batch_size = ctx.batch_size;
-            // Held only on the non-fused streaming path; the fused path
-            // consumes its sender inside `merge_fused_interleave`.
-            let mut nonfused_sender: Option<
-                crossbeam_channel::Sender<crate::executor::stream_event::StreamEvent>,
-            > = None;
-            let merged: Vec<(Record, u64)> = if fused_mode {
-                let handoff = match (streaming_sender, merge_charge.as_ref()) {
-                    (Some(sender), Some(charge)) => Some(MergeStreamHandoff {
-                        sender,
-                        charge,
-                        batch_size: merge_batch_size,
-                    }),
-                    _ => None,
-                };
-                merge_fused_interleave(
-                    ctx,
-                    current_dag,
-                    name,
-                    &sorted_preds,
-                    merge_output_schema.as_ref(),
-                    handoff,
-                )?
-            } else {
-                nonfused_sender = streaming_sender;
-                let total: usize = sorted_preds
-                    .iter()
-                    .map(|p| ctx.node_buffers.get(p).map_or(0, |b| b.len_hint()))
-                    .sum();
-                let mut merged = Vec::with_capacity(total);
-                let emit = |merged: &mut Vec<(Record, u64)>,
-                            upstream_name: &str,
-                            mut record: Record,
-                            rn: u64|
-                 -> Result<(), PipelineError> {
-                    if let Some(canonical) = merge_output_schema.as_ref() {
-                        check_input_schema(
-                            canonical,
-                            record.schema(),
-                            name,
-                            "merge",
-                            upstream_name,
-                        )?;
-                        // Rebuild record with the Merge's
-                        // `Arc<Schema>` so downstream operators hit
-                        // the ptr_eq fast path regardless of which
-                        // input the record originated from. Carry the
-                        // per-record envelope context across the
-                        // rebuild — each merged row keeps the document
-                        // it came from.
-                        let doc_ctx = Arc::clone(record.doc_ctx());
-                        let values = record.values().to_vec();
-                        record = Record::new(Arc::clone(canonical), values);
-                        record.set_doc_ctx(doc_ctx);
-                    }
-                    merged.push((record, rn));
-                    Ok(())
-                };
-
-                // Every input's punctuations accumulate in the outer
-                // `all_puncts`; the post-merge dedup pass collapses
-                // multiple-input duplicates so each `DocumentOpen` and
-                // `DocumentClose` traverses downstream exactly once.
-                match mode {
-                    crate::config::MergeMode::Concat => {
-                        for pred in &sorted_preds {
-                            let upstream_name = current_dag.graph[*pred].name().to_string();
-                            if let Some(buf) = drain_node_buffer_slot(ctx, *pred) {
-                                for event in buf.drain() {
-                                    match event? {
-                                        crate::executor::stream_event::StreamEvent::Record(
-                                            record,
-                                            rn,
-                                        ) => {
-                                            emit(&mut merged, &upstream_name, record, rn)?;
-                                        }
-                                        crate::executor::stream_event::StreamEvent::Punctuation(
-                                            p,
-                                        ) => {
-                                            all_puncts.push(p);
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    crate::config::MergeMode::Interleave => {
-                        // Round-robin across pre-buffered predecessor
-                        // outputs. Reached only when predecessors are
-                        // not all Sources (the fused path took
-                        // those); the inputs are produced by upstream
-                        // operator arms that wrote to `node_buffers`.
-                        use std::collections::VecDeque;
-                        let upstream_names: Vec<String> = sorted_preds
-                            .iter()
-                            .map(|p| current_dag.graph[*p].name().to_string())
-                            .collect();
-                        let mut deques: Vec<VecDeque<(Record, u64)>> = sorted_preds
-                            .iter()
-                            .map(|pred| -> Result<VecDeque<(Record, u64)>, PipelineError> {
-                                match drain_node_buffer_slot(ctx, *pred) {
-                                    Some(nb) => {
-                                        let (records, puncts) = nb.drain_split()?;
-                                        all_puncts.extend(puncts);
-                                        Ok(VecDeque::from(records))
-                                    }
-                                    None => Ok(VecDeque::new()),
-                                }
-                            })
-                            .collect::<Result<Vec<_>, _>>()?;
-                        let mut rng = interleave_seed.map(fastrand::Rng::with_seed);
-                        let mut cursor = 0usize;
-                        loop {
-                            let ready: Vec<usize> = deques
-                                .iter()
-                                .enumerate()
-                                .filter_map(|(i, d)| (!d.is_empty()).then_some(i))
-                                .collect();
-                            if ready.is_empty() {
-                                break;
-                            }
-                            let pick_idx = match rng.as_mut() {
-                                Some(r) => ready[r.usize(0..ready.len())],
-                                None => {
-                                    let mut chosen = ready[0];
-                                    for _ in 0..deques.len() {
-                                        let candidate = cursor % deques.len();
-                                        cursor = cursor.wrapping_add(1);
-                                        if !deques[candidate].is_empty() {
-                                            chosen = candidate;
-                                            break;
-                                        }
-                                    }
-                                    chosen
-                                }
-                            };
-                            if let Some((record, rn)) = deques[pick_idx].pop_front() {
-                                emit(&mut merged, &upstream_names[pick_idx], record, rn)?;
-                            }
-                        }
-                    }
-                }
-                merged
-            };
-
-            // Merge is rejected as a window root at compile time
-            // (E150d) because the multi-producer concatenation has no
-            // single producer identity for record_pos to anchor
-            // against. The call below is a defense-in-depth no-op:
-            // the helper iterates plan.indices_to_build and matches
-            // nothing because no spec roots at a Merge NodeIndex.
-            // Punctuation dedup pass: emit `DocumentOpen` once per
-            // document (first sighting wins) and `DocumentClose` only
-            // after every input branch has closed the same document.
-            let mut deduped_puncts: Vec<crate::executor::stream_event::Punctuation> = Vec::new();
-            let mut open_seen: std::collections::HashSet<clinker_record::DocumentId> =
-                std::collections::HashSet::new();
-            let mut close_counts: std::collections::HashMap<clinker_record::DocumentId, usize> =
-                std::collections::HashMap::new();
-            for p in all_puncts {
-                let id = p.doc_id();
-                match p.kind() {
-                    crate::executor::stream_event::PunctuationKind::DocumentOpen => {
-                        if open_seen.insert(id) {
-                            deduped_puncts.push(p);
-                        }
-                    }
-                    crate::executor::stream_event::PunctuationKind::DocumentClose => {
-                        let count = close_counts.entry(id).or_insert(0);
-                        *count += 1;
-                        if *count == fan_in_degree {
-                            deduped_puncts.push(p);
-                        }
-                    }
-                }
-            }
-
-            // Non-fused streaming path: the predecessors' slots are
-            // already drained into the full `merged` Vec, so this does not
-            // shrink the Merge's own working set; what it saves is the
-            // second copy — hand `merged` straight to the downstream
-            // Output thread over the bounded channel rather than admitting
-            // a charged `node_buffers` slot the Output would re-drain, and
-            // overlap the writer with the next topo node. (The fused
-            // Merge.interleave path is the true one-batch streamer; it
-            // forwards records off the live Source channels inside
-            // `merge_fused_interleave` and returns an empty `merged`, its
-            // sender already dropped there.) The eligibility predicate
-            // certified this Merge roots no window and tees to no deferred
-            // region, so the helper calls below are correctly skipped.
-            // Dropping `nonfused_sender` at the end of this branch
-            // disconnects the writer thread's `recv` loop.
-            if let Some(sender) = nonfused_sender {
-                let charge = merge_charge
-                    .as_ref()
-                    .expect("streaming sender implies a registered charge handle");
-                stream_linear_producer_emit(
-                    &sender,
-                    merge_batch_size,
-                    name,
-                    merged,
-                    deduped_puncts,
-                    charge,
-                )?;
-                return Ok(());
-            }
-
-            finalize_node_rooted_windows(ctx, current_dag, node_idx, &merged)?;
-            tee_emit_to_region_input_buffers(ctx, current_dag, node_idx, &merged)?;
-            let nb = admit_node_buffer(
-                ctx,
-                name,
-                node_idx,
-                merged,
-                deduped_puncts,
-                node_buffer_spill_allowed(current_dag, node_idx),
-            )?;
-            ctx.node_buffers.insert(node_idx, nb);
+        PlanNode::Merge { .. } => {
+            crate::executor::merge_dispatch::dispatch_merge(ctx, current_dag, node_idx, &node)?;
         }
 
-        PlanNode::Sort {
-            ref name,
-            ref sort_fields,
-            ..
-        } => {
-            // Enforcer-sort dispatch. Carries `row_num` through
-            // the sort permutation as the `SortBuffer<u64>`
-            // payload — the Record itself carries every field
-            // value, emitted content, and metadata, so no
-            // parallel bookkeeping map rides alongside.
-            use crate::pipeline::sort_buffer::{SortBuffer, SortedOutput};
-
-            let predecessors: Vec<NodeIndex> = current_dag
-                .graph
-                .neighbors_directed(node_idx, Direction::Incoming)
-                .collect();
-            let (input_records, input_puncts): (
-                Vec<(Record, u64)>,
-                Vec<crate::executor::stream_event::Punctuation>,
-            ) = match predecessors
-                .iter()
-                .find_map(|p| drain_node_buffer_slot(ctx, *p))
-            {
-                Some(nb) => nb.drain_split()?,
-                None => (Vec::new(), Vec::new()),
-            };
-
-            if input_records.is_empty() {
-                tee_emit_to_region_input_buffers(ctx, current_dag, node_idx, &[])?;
-                // An empty input still registers a (zero-byte) consumer
-                // via `admit_node_buffer` for symmetry with the non-empty
-                // path, so the arbitrator's pull-mode registry treats
-                // every Sort insert uniformly.
-                let nb = admit_node_buffer(
-                    ctx,
-                    name,
-                    node_idx,
-                    Vec::new(),
-                    input_puncts,
-                    node_buffer_spill_allowed(current_dag, node_idx),
-                )?;
-                ctx.node_buffers.insert(node_idx, nb);
-                return Ok(());
-            }
-
-            let schema = input_records[0].0.schema().clone();
-            let mem_limit = parse_memory_limit(ctx.config);
-            let mut buf: SortBuffer<u64> = SortBuffer::new(
-                sort_fields.clone(),
-                mem_limit,
-                Some(ctx.spill_root_path.to_path_buf()),
-                schema,
-            );
-
-            let sort_timer = stage_metrics::StageTimer::new(stage_metrics::StageName::Sort);
-            let sort_count = input_records.len() as u64;
-            // CPU-bound: sort buffer push + per-batch comparison + spill
-            // I/O. The kernel owns its input (`input_records`, `buf`) and
-            // borrows nothing from `ctx`, so it runs on the shared Rayon
-            // pool; output order is fully determined by the sort itself,
-            // so pool scheduling cannot perturb it.
-            let sorted = ctx.kernel_pool.install(|| {
-                for (record, row_num) in input_records {
-                    buf.push(record, row_num);
-                    if buf.should_spill() {
-                        buf.sort_and_spill().map_err(|e| {
-                            PipelineError::Io(std::io::Error::other(format!(
-                                "sort enforcer '{name}' spill failed: {e}"
-                            )))
-                        })?;
-                    }
-                }
-                buf.finish().map_err(|e| {
-                    PipelineError::Io(std::io::Error::other(format!(
-                        "sort enforcer '{name}' finish failed: {e}"
-                    )))
-                })
-            })?;
-
-            let mut out: Vec<(Record, u64)> = Vec::with_capacity(sort_count as usize);
-            match sorted {
-                SortedOutput::InMemory(pairs) => {
-                    for (record, row_num) in pairs {
-                        out.push((record, row_num));
-                    }
-                }
-                SortedOutput::Spilled(files) => {
-                    // Spill files are individually sorted but not
-                    // globally merged. For correctness when multiple
-                    // spill files exist, a k-way merge is required.
-                    // Single-file fast path is exact.
-                    if files.len() > 1 {
-                        return Err(PipelineError::Io(std::io::Error::other(format!(
-                            "sort enforcer '{name}' produced {} spill files; \
-                             k-way merge for enforcer sort is not yet implemented \
-                             (memory.limit too small for input)",
-                            files.len()
-                        ))));
-                    }
-                    for file in files {
-                        let reader = file.reader().map_err(|e| {
-                            PipelineError::Io(std::io::Error::other(format!(
-                                "sort enforcer '{name}' spill read failed: {e}"
-                            )))
-                        })?;
-                        for entry in reader {
-                            let (record, row_num) = entry.map_err(|e| {
-                                PipelineError::Io(std::io::Error::other(format!(
-                                    "sort enforcer '{name}' spill decode failed: {e}"
-                                )))
-                            })?;
-                            out.push((record, row_num));
-                        }
-                    }
-                }
-            }
-            ctx.collector
-                .record(sort_timer.finish(sort_count, sort_count));
-            tee_emit_to_region_input_buffers(ctx, current_dag, node_idx, &out)?;
-            let nb = admit_node_buffer(
-                ctx,
-                name,
-                node_idx,
-                out,
-                input_puncts,
-                node_buffer_spill_allowed(current_dag, node_idx),
-            )?;
-            ctx.node_buffers.insert(node_idx, nb);
+        PlanNode::Sort { .. } => {
+            crate::executor::sort_dispatch::dispatch_sort(ctx, current_dag, node_idx, &node)?;
         }
 
         PlanNode::Aggregation { .. } => {
@@ -4107,7 +3332,7 @@ pub(crate) fn dispatch_plan_node(
         }
 
         PlanNode::CorrelationCommit { .. } => {
-            crate::executor::commit::orchestrate(ctx, current_dag)?;
+            crate::executor::correlation_dispatch::dispatch_correlation_commit(ctx, current_dag)?;
         }
     }
 
@@ -4172,374 +3397,6 @@ pub(crate) struct CorrelationErrorRecord {
     pub(crate) error_message: String,
     pub(crate) stage: Option<String>,
     pub(crate) route: Option<String>,
-}
-
-/// Walk every correlation buffer and emit the per-group commit:
-/// flush-to-writer for clean groups, DLQ-with-trigger for dirty groups,
-/// overflow-aware DLQ for groups that tripped `max_group_buffer`.
-///
-/// Two-phase: phase 1 walks every group once, accumulating clean
-/// records into a per-output queue and emitting DLQ entries for
-/// dirty/overflow groups. Phase 2 opens each writer ONCE, writes the
-/// accumulated queue, and flushes — one writer take per Output, not
-/// one per group. The phasing is necessary because every group flush
-/// would otherwise contend for `ctx.writers.remove(name)`, leaving
-/// subsequent groups silently dropped.
-///
-/// Single source of truth for both strict and relaxed pipelines. The
-/// relaxed-CK orchestrator runs its cascading-retraction loop, then
-/// delegates here once the post-recompute aggregate output rows have
-/// been substituted into the buffer. The per-group emission shape is
-/// identical; only the path that populates the buffer differs.
-pub(crate) fn commit_correlation_buffers(
-    ctx: &mut ExecutorContext<'_>,
-) -> Result<(), PipelineError> {
-    use std::collections::BTreeMap;
-
-    let Some(mut buffers) = ctx.correlation_buffers.take() else {
-        return Ok(());
-    };
-
-    // Stable iteration order — sort by formatted group key so DLQ
-    // emission and writer flush are deterministic across runs.
-    let mut group_keys: Vec<Vec<GroupByKey>> = buffers.keys().cloned().collect();
-    group_keys.sort_by_key(|k| format_group_key(k));
-
-    // Phase 1: per-group disposition. Clean groups' records flow
-    // into `clean_per_output` keyed by output_name; the BTreeMap key
-    // ordering is the writer-flush order. Dirty groups land in the
-    // DLQ inline.
-    let mut clean_per_output: BTreeMap<String, Vec<CorrelationRecordSlot>> = BTreeMap::new();
-    for group_key in group_keys {
-        let Some(group) = buffers.remove(&group_key) else {
-            continue;
-        };
-        commit_one_group(ctx, &group_key, group, &mut clean_per_output)?;
-    }
-
-    // Phase 2: drain the clean per-output queues to writers.
-    flush_clean_records_to_writers(ctx, clean_per_output)?;
-
-    Ok(())
-}
-
-fn format_group_key(key: &[GroupByKey]) -> String {
-    let parts: Vec<String> = key
-        .iter()
-        .map(|k| match k {
-            GroupByKey::Null => "null".to_string(),
-            GroupByKey::Bool(b) => b.to_string(),
-            GroupByKey::Int(i) => i.to_string(),
-            GroupByKey::Float(bits) => f64::from_bits(*bits).to_string(),
-            GroupByKey::Str(s) => format!("{s:?}"),
-            GroupByKey::Date(d) => d.to_string(),
-            GroupByKey::DateTime(ts) => ts.to_string(),
-        })
-        .collect();
-    format!("[{}]", parts.join(", "))
-}
-
-fn commit_one_group(
-    ctx: &mut ExecutorContext<'_>,
-    group_key: &[GroupByKey],
-    group: CorrelationGroupBuffer,
-    clean_per_output: &mut std::collections::BTreeMap<String, Vec<CorrelationRecordSlot>>,
-) -> Result<(), PipelineError> {
-    let CorrelationGroupBuffer {
-        records,
-        error_rows,
-        error_messages,
-        total_records,
-        overflowed,
-    } = group;
-
-    if overflowed {
-        // Overflow disposition: one root-cause DLQ entry with
-        // category=GroupSizeExceeded and trigger=true; every other
-        // buffered record of the group becomes a collateral with
-        // category=Correlated and trigger=false. Per-record original
-        // records come from `records` (the projected buffer holds the
-        // un-projected original alongside).
-        let mut emitted_trigger = false;
-        let group_repr = format_group_key(group_key);
-        let overflow_msg = PipelineError::CorrelationGroupOverflow {
-            group_key: group_repr.clone(),
-            count: total_records,
-        }
-        .to_string();
-        // Dedup distinct rows by (source, row_num) so Route fan-out
-        // (one row to N outputs) emits one DLQ entry, not N. Pairing on
-        // `source_name` matters under multi-source ingest where row_num
-        // is per-source.
-        let mut seen_rows: HashSet<(Arc<str>, u64)> = HashSet::new();
-        for slot in &records {
-            let source_name = source_name_arc_of(&slot.original_record);
-            if !seen_rows.insert((Arc::clone(&source_name), slot.row_num)) {
-                continue;
-            }
-            let (category, trigger, error_message) = if !emitted_trigger {
-                emitted_trigger = true;
-                (
-                    crate::dlq::DlqErrorCategory::GroupSizeExceeded,
-                    true,
-                    overflow_msg.clone(),
-                )
-            } else {
-                (
-                    crate::dlq::DlqErrorCategory::Correlated,
-                    false,
-                    format!("correlated with failure in group: {overflow_msg}"),
-                )
-            };
-            push_dlq(
-                ctx,
-                DlqEntry {
-                    source_row: slot.row_num,
-                    category,
-                    error_message,
-                    original_record: slot.original_record.clone(),
-                    stage: Some("correlation_commit".to_string()),
-                    route: None,
-                    trigger,
-                    source_name,
-                    triggering_field: None,
-                    triggering_value: None,
-                },
-            )?;
-        }
-        // Per-source rollback rewind: each contributing source rewinds
-        // its `rollback_cursors` entry to the lowest `row_num` of any
-        // group member from that source. The cursor narrows the replay
-        // anchor so a downstream resume reprocesses every record that
-        // contributed to the overflowing group, including those whose
-        // forward operators had already advanced the cursor past them.
-        // No causal-source attribution is required for an overflow —
-        // every contributing source shared blame proportionally — so
-        // every source contributing a slot rewinds independently.
-        let mut per_source_min: HashMap<Arc<str>, u64> = HashMap::new();
-        for slot in &records {
-            let sn = source_name_arc_of(&slot.original_record);
-            per_source_min
-                .entry(sn)
-                .and_modify(|m| *m = (*m).min(slot.row_num))
-                .or_insert(slot.row_num);
-        }
-        for (sn, min_rn) in per_source_min {
-            ctx.rollback_cursors
-                .entry(sn)
-                .and_modify(|c| *c = (*c).min(min_rn))
-                .or_insert(min_rn);
-        }
-        return Ok(());
-    }
-
-    let group_dirty = !error_rows.is_empty();
-    if !group_dirty {
-        // Clean group → drop the records into the per-output queue
-        // for batched flush after every group has been visited.
-        for slot in records {
-            clean_per_output
-                .entry(slot.output_name.clone())
-                .or_default()
-                .push(slot);
-        }
-        return Ok(());
-    }
-
-    // Dirty group → drop projected records, emit DLQ entries for every
-    // distinct row_num touched by the group. Triggers come from
-    // `error_messages`; collaterals come from `records` (rows that
-    // succeeded their leg but get rolled back because the group failed).
-    let first_err_message = error_messages
-        .first()
-        .map(|e| e.error_message.clone())
-        .unwrap_or_else(|| "unknown".to_string());
-
-    // Per-source narrowing: a collateral slot is spared whenever its
-    // originating Source did not contribute any trigger error to this
-    // group. With multi-source ingest, a single trigger from `src_b`
-    // would otherwise DLQ every co-grouped `src_a` slot purely by
-    // sharing the correlation key — `src_a` had no causal role in the
-    // failure. The set is built from `error_messages` before the
-    // trigger loop so each trigger's `$source.name` stamp drives its
-    // own narrowing of the collateral walk below. Empty in a
-    // single-source pipeline by construction: every co-grouped slot
-    // shares the failing source, so the wider behavior is
-    // bit-identical to today's pipeline-wide collateral DLQ.
-    let failing_sources: HashSet<Arc<str>> = error_messages
-        .iter()
-        .map(|err| source_name_arc_of(&err.original_record))
-        .collect();
-
-    // Dedup distinct rows by (source, row_num) so Route fan-out (one
-    // row to N outputs) emits one DLQ entry, not N. Pairing on
-    // `source_name` matters under multi-source ingest where row_num is
-    // per-source — pre-sprint-7 `HashSet<u64>` collided across sources
-    // and silently dropped co-rowed slots from the collateral walk.
-    let mut seen_rows: HashSet<(Arc<str>, u64)> = HashSet::new();
-    // Trigger entries: one per distinct erroring row. Multiple errors
-    // per row (different branches failing) collapse to a single trigger
-    // entry carrying the first error.
-    for err in &error_messages {
-        let source_name = source_name_arc_of(&err.original_record);
-        if !seen_rows.insert((Arc::clone(&source_name), err.row_num)) {
-            continue;
-        }
-        push_dlq(
-            ctx,
-            DlqEntry {
-                source_row: err.row_num,
-                category: err.category,
-                error_message: err.error_message.clone(),
-                original_record: err.original_record.clone(),
-                stage: err.stage.clone(),
-                route: err.route.clone(),
-                trigger: true,
-                source_name,
-                triggering_field: None,
-                triggering_value: None,
-            },
-        )?;
-    }
-    // Collateral entries: every other distinct row that flowed through
-    // the group's Output buffers but didn't itself error. Two sparing
-    // axes apply, in order:
-    //   1. Per-source narrowing — a slot from a non-failing source is
-    //      always spared. Falls back to today's `Any` semantics within
-    //      single-source pipelines.
-    //   2. CorrelationFanoutPolicy interpretation — Any/All/Primary
-    //      operate WITHIN the failing-source's records.
-    // The resolved policy is read per-Output (per-Combine / per-Output
-    // overrides win against the pipeline default). Today the strict
-    // path always resolves to `Any` so axis 2 sparing is a no-op; the
-    // wire is in place for the relaxed-CK orchestrator to substitute a
-    // different policy.
-    for slot in &records {
-        let slot_source = source_name_arc_of(&slot.original_record);
-        if seen_rows.contains(&(Arc::clone(&slot_source), slot.row_num)) {
-            continue;
-        }
-        // Per-source spare: slot's Source did not contribute a trigger
-        // to this group, so it never had a causal role in the failure.
-        // Slots stamped as `MERGED_SOURCE_NAME` carry no single-source
-        // attribution — Combine outputs (multi-input fold) and
-        // synthetic aggregate emits both reach this branch. Treating
-        // them as ambiguous, they stay on the collateral path so a
-        // downstream failure on a Combine-derived row still flushes
-        // the upstream-trigger correlation under today's
-        // pipeline-wide-equivalent semantics. Per-source narrowing
-        // only spares slots whose stamp identifies a distinct
-        // non-failing Source.
-        let slot_attributable = slot_source.as_ref() != MERGED_SOURCE_NAME.as_ref();
-        if slot_attributable && !failing_sources.contains(&slot_source) {
-            clean_per_output
-                .entry(slot.output_name.clone())
-                .or_default()
-                .push(slot.clone());
-            continue;
-        }
-        let policy = crate::executor::commit::output_fanout_policy(ctx, &slot.output_name);
-        // Strict path: every slot's CK identity already equals the
-        // group's CK identity by construction (the buffer key is
-        // shared). `is_full_tuple_match` is `true`; `is_primary_match`
-        // is also `true`. Under `Any` the slot DLQs; under `All` the
-        // slot DLQs (full match); under `Primary` the slot DLQs
-        // (primary match). The relaxed orchestrator's flush path
-        // populates these flags differently per slot when multi-CK
-        // fan-out makes the match partial.
-        let spare = crate::executor::commit::should_spare_collateral(policy, true, true);
-        if spare {
-            // Spared collateral lands in the per-output clean queue
-            // alongside the originally-clean records.
-            clean_per_output
-                .entry(slot.output_name.clone())
-                .or_default()
-                .push(slot.clone());
-            continue;
-        }
-        if !seen_rows.insert((Arc::clone(&slot_source), slot.row_num)) {
-            continue;
-        }
-        push_dlq(
-            ctx,
-            DlqEntry {
-                source_row: slot.row_num,
-                category: crate::dlq::DlqErrorCategory::Correlated,
-                error_message: format!("correlated with failure in group: {first_err_message}"),
-                original_record: slot.original_record.clone(),
-                stage: Some("correlation_commit".to_string()),
-                route: None,
-                trigger: false,
-                source_name: slot_source,
-                triggering_field: None,
-                triggering_value: None,
-            },
-        )?;
-    }
-
-    Ok(())
-}
-
-fn flush_clean_records_to_writers(
-    ctx: &mut ExecutorContext<'_>,
-    per_output: std::collections::BTreeMap<String, Vec<CorrelationRecordSlot>>,
-) -> Result<(), PipelineError> {
-    for (output_name, slots) in per_output {
-        if slots.is_empty() {
-            continue;
-        }
-        let out_cfg = ctx
-            .output_configs
-            .iter()
-            .find(|o| o.name == output_name)
-            .unwrap_or(ctx.primary_output);
-        let output_schema = Arc::clone(slots[0].projected.schema());
-
-        let Some(raw_writer) = ctx.writers.remove(&output_name) else {
-            continue;
-        };
-        match build_format_writer(out_cfg, raw_writer, Arc::clone(&output_schema)) {
-            Ok(mut writer) => {
-                let mut write_failed = false;
-                for slot in &slots {
-                    let write_result = {
-                        let _guard = ctx.write_timer.guard();
-                        writer.write_record(&slot.projected)
-                    };
-                    if let Err(e) = write_result {
-                        ctx.output_errors.push(PipelineError::from(e));
-                        write_failed = true;
-                        break;
-                    }
-                }
-                if !write_failed {
-                    let flush_result = {
-                        let _guard = ctx.write_timer.guard();
-                        writer.flush()
-                    };
-                    if let Err(e) = flush_result {
-                        ctx.output_errors.push(PipelineError::from(e));
-                    }
-                }
-            }
-            Err(e) => ctx.output_errors.push(e),
-        }
-
-        // Counter accounting: per-row newly-ok bookkeeping was
-        // deferred at Output arm time when buffering. Apply it here
-        // so records never count toward `ok_count` if their group
-        // rolled back.
-        let mut newly_ok: u64 = 0;
-        for slot in &slots {
-            if ctx.ok_source_rows.insert(slot.row_num) {
-                newly_ok += 1;
-            }
-        }
-        ctx.counters.ok_count += newly_ok;
-        ctx.counters.records_written += slots.len() as u64;
-        ctx.records_emitted += slots.len() as u64;
-    }
-    Ok(())
 }
 
 /// Collect parent-scope records keyed by composition input port name.
