+++ crates/clinker-core/src/executor/mod.rs
@@ -5,10 +5,14 @@ pub(crate) mod batch_handoff;
 pub mod combine;
 pub(crate) mod combine_dispatch;
 pub(crate) mod commit;
+pub(crate) mod correlation_dispatch;
 pub(crate) mod dispatch;
+pub(crate) mod merge_dispatch;
 pub mod node_buffer;
 pub(crate) mod node_buffer_spill;
+pub(crate) mod route_dispatch;
 mod schema_check;
+pub(crate) mod sort_dispatch;
 pub mod source_stream;
 pub(crate) mod stream_event;
 pub(crate) mod time_window;
